pipe_sequence traversing was refactored  into towo different functions, and new fle creating to store them
